+++ src/Pages/AboutBook.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import TopNavbar from '../../src/components/TopNavbar';
-import AboutBook from '../components/AboutBook';
-import Footer from '../../src/components/Footer';
+import TopNavbar from '../components/TopNavbar';
+import AboutBook from '../components/BookDetails';
+import Footer from '../components/Footer';
 
 const AboutBookPage = ({book}) => {
   return (
