+++ src/Pages/BookSearch.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import useBookSearchResults from '../hooks/useBookSearchResults';
 import '../css/BookSearchPage.css';
 
-const BookSearchPage = () => {
+const BookSearch = () => {
     const [searchTerm, setSearchTerm] = useState('');
     const { books, isLoading, isError } = useBookSearchResults(searchTerm);
     const [isModalOpen, setIsModalOpen] = useState(false);
@@ -48,4 +48,4 @@ const BookSearchPage = () => {
     );
 };
 
-export default BookSearchPage;
+export default BookSearch;
