+++ src/Pages/Home.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { useBooks } from '../hooks/useBooks'
-import BookList from '../../src/components/BookList';
-import TopNavbar from '../../src/components/TopNavbar'; // Import the TopNavbar component
-import Greeting from '../../src/components/Greeting'; // Import the Greeting component
-import QrReader from '../../src/components/QrReader';
-import Footer from '../../src/components/Footer'; // Import the Footer component
+import BookList from '../components/BookList';
+import TopNavbar from '../components/TopNavbar'; // Import the TopNavbar component
+import Greeting from '../components/Greeting'; // Import the Greeting component
+import QrReader from '../components/QrReader';
+import Footer from '../components/Footer'; // Import the Footer component
 import { userData } from '../helpers/userStorage';
 
-const StudentHomePage = () => {
+const Home = () => {
   const {books, updateBooks} = useBooks();
   const {username} = userData() || {}
   
@@ -27,4 +27,4 @@ const StudentHomePage = () => {
   );
 };
 
-export default StudentHomePage;
+export default Home;
